Improved the Recent Visits page blacklist.
Rewrote the script so all the checklists from any observer can be
hidden by clicking on a checkmark next to their name. The checklist
are only hidden until the page is loaded.

Updated the list of URLs the script work for to include hotspots.
Previously it only worked for regions (contrary to the descirption).