Added script to restore the color to visited links.
The link color for links to visited checklist pages is the same blue as
unvisited ones. That makes it impossible, when reviewing all the checklists,
to see which ones you have already read. This is a particular problem if
you read the recents visits page on a regular basis.

Now the color for visited links to checklists is set to dark organge. This
is a little more obvious that the default purple colour (used at least in
FireFox).